# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VNDRLY Field Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/1099-transmitter-settings:
    get:
      operationId: getFireTransmitterSettings
      tags: [platform]
      summary: Get IRS FIRE transmitter settings (singleton)
      description: |
        Admin-only. Returns the singleton fire_transmitter_settings row
        used as the source of truth for the T record on every IRS 1099
        FIRE submission. When no row has ever been saved every value is
        returned as null and real (non-test) FIRE downloads are
        blocked. Includes a `missing` array listing required fields
        that are still blank, mirroring the structured guard the
        e-file route enforces.
      responses:
        "200":
          description: Transmitter settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireTransmitterSettings"
        "403":
          description: Admin role required
    put:
      operationId: updateFireTransmitterSettings
      tags: [platform]
      summary: Save IRS FIRE transmitter settings
      description: |
        Admin-only. Replaces every column on the singleton row in one
        write and appends an audit row capturing the diff and the
        actor. Performs the same validation the FIRE e-file route
        enforces — every field is required and the address must parse
        into city/state/zip — so a successful save means a real
        (non-test) FIRE submission can immediately be generated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFireTransmitterSettingsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FireTransmitterSettings"
        "400":
          description: Validation failed
        "403":
          description: Admin role required
  /admin/1099-transmitter-settings/history:
    get:
      operationId: listFireTransmitterSettingsHistory
      tags: [platform]
      summary: List recent IRS FIRE transmitter setting changes (admin only)
      description: |
        Admin-only history view over `fire_transmitter_settings_audit_log`.
        One row per save that actually changed at least one column,
        newest first. Each entry carries the actor (display name + email
        joined from `users` at read time, falling back to NULL if the
        user has since been deleted), the actor role / IP / user-agent
        captured at save time, and the per-field `{ before, after }`
        diff jsonb so the UI can render "admin X changed our IRS TCC
        from … to … on …" without re-deriving state.

        Capped at `limit` rows (default 50, max 200) so the page can
        be paged with `offset`.
      parameters:
        - name: limit
          in: query
          required: false
          description: Max rows to return. Default 50, capped at 200.
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          required: false
          description: Skip this many rows before returning. Default 0.
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Paged list of transmitter audit rows (newest first)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFireTransmitterSettingsHistoryResponse"
        "403":
          description: Admin role required

  # ── Demo-account picker labels (per locale, admin-editable) ──
...
components:
  schemas:
...
    FireTransmitterSettings:
      type: object
      description: |
        Singleton row of IRS FIRE transmitter info. Each value is null
        when not yet saved (UI shows it as blank); a save replaces all
        seven values together.
      properties:
        tcc:
          type: ["string", "null"]
          description: 5-character Transmitter Control Code assigned by the IRS.
        ein:
          type: ["string", "null"]
          description: Transmitter EIN (9 digits, no dashes).
        name:
          type: ["string", "null"]
          description: Transmitter legal name (40 chars, written into the T record).
        address:
          type: ["string", "null"]
          description: |
            Single-line transmitter mailing address — the FIRE
            generator parses it into street/city/state/zip.
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
          description: Digits only, 10-character US phone number.
        updatedAt:
          type: ["string", "null"]
          format: date-time
        updatedByUserId:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
          description: |
            Display name of the admin who last saved the row, joined
            from `users.display_name`. Null when no row has been saved
            yet or when the linked user has been deleted.
        updatedByEmail:
          type: ["string", "null"]
          description: |
            Email of the admin who last saved the row, used as a
            fallback in the UI when `updatedByName` is blank.
        missing:
          type: array
          items:
            type: string
          description: |
            Field names that are still blank/unparseable and would
            block a real (non-test) FIRE submission. Empty when every
            required value is set and the address parses cleanly.
      required:
        - tcc
        - ein
        - name
        - address
        - contactName
        - contactEmail
        - contactPhone
        - updatedAt
        - updatedByUserId
        - updatedByName
        - updatedByEmail
        - missing
    UpdateFireTransmitterSettingsBody:
      type: object
      description: |
        Replaces every column on the singleton transmitter row. All
        seven fields are required and validated by the same rules the
        e-file route enforces (non-blank; address must parse into
        street/city/state/zip).
      properties:
        tcc:
          type: string
          minLength: 1
        ein:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        contactName:
          type: string
          minLength: 1
        contactEmail:
          type: string
          minLength: 1
        contactPhone:
          type: string
          minLength: 1
      required:
        - tcc
        - ein
        - name
        - address
        - contactName
        - contactEmail
        - contactPhone
    FireTransmitterSettingsHistoryFieldChange:
      type: object
      description: |
        Per-field `{ before, after }` diff entry. Both sides are
        rendered as strings in the UI; `before` is null when the
        column had never been saved before this change.
      properties:
        before:
          type: ["string", "null"]
        after:
          type: ["string", "null"]
      required: [before, after]
    FireTransmitterSettingsHistoryEntry:
      type: object
      description: |
        One row from `fire_transmitter_settings_audit_log` joined to
        `users` for actor display name + email at read time. Each row
        is a single save that changed at least one column.
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        changes:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FireTransmitterSettingsHistoryFieldChange"
          description: |
            Map of changed-column-name → `{ before, after }`. Only
            columns whose value actually changed in this save appear.
        actorUserId:
          type: ["integer", "null"]
        actorDisplayName:
          type: ["string", "null"]
          description: |
            `users.display_name` of the actor at read time. NULL if
            the user has since been deleted (the FK is
            `ON DELETE SET NULL`).
        actorEmail:
          type: ["string", "null"]
        actorRole:
          type: string
        actorIp:
          type: ["string", "null"]
        actorUserAgent:
          type: ["string", "null"]
      required: [id, createdAt, changes, actorRole]
    ListFireTransmitterSettingsHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FireTransmitterSettingsHistoryEntry"
        total:
          type: integer
          description: |
            Total number of rows in
            `fire_transmitter_settings_audit_log` (independent of the
            current `limit` / `offset`).
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
...